+++ ChangeLog.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - *OpenAI* client (work in progress)
-- functions `get-header`, `get-parameter`, and `get-long-parameter` to :ring module
+- functions `get-request-parameter`, and `get-request-long-parameter` to 
+  :ring-util module